Add sunspot_solr as a dependency.
The primary role of this gem is to provide a Solr that we can run tests
againsts. The sunspot_rails gem does not depend on sunspot_solr as it
provides a lot of functionality other than running a server. That said,
it does depend on the gem in order to use `Sunspot::Rails::Server`.
Since we have a hard dependency on that, we need to also make sure the
underlying dependency of sunspot_solr is included.